# packages/openapi-generator/tmp/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/posts": {
      "post": {
        "operationId": "PostController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "content": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PostContentSchema"
                    }
                  },
                  "coverImage": {
                    "type": "string",
                    "format": "uri"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": ["title", "content"],
                "title": "CreatePostSchema",
                "description": "Schema for creating/updating a post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostSchema"
                }
              }
            }
          }
        },
        "tags": ["Admin Posts"]
      },
      "get": {
        "operationId": "PostController_getUserPosts",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Starting position of the query",
            "schema": {
              "default": 0,
              "description": "Starting position of the query",
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items to return",
            "schema": {
              "default": 20,
              "description": "Number of items to return",
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Schema for sorting items",
            "schema": {
              "$ref": "#/components/schemas/PostController_getUserPosts_sortArray"
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filtering query string, in the format of \"property:rule[:value];property:rule[:value];...\"\nAvailable rules: eq, neq, gt, gte, lt, lte, like, nlike, in, nin, isnull, isnotnull \nAvailable properties: title, tag",
            "schema": {
              "$ref": "#/components/schemas/PostController_getUserPosts_filterArray"
            },
            "style": "form",
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostsSchema"
                }
              }
            }
          }
        },
        "tags": ["Admin Posts"]
      }
    },
    "/api/admin/posts/{id}": {
      "put": {
        "operationId": "PostController_updatePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "content": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PostContentSchema"
                    }
                  },
                  "coverImage": {
                    "type": "string",
                    "format": "uri"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "title": "UpdatePostSchema",
                "description": "Schema for updating a post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostSchema"
                }
              }
            }
          }
        },
        "tags": ["Admin Posts"]
      },
      "get": {
        "operationId": "PostController_getUserPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostSchema"
                }
              }
            }
          }
        },
        "tags": ["Admin Posts"]
      }
    },
    "/api/admin/posts/{id}/publish": {
      "patch": {
        "operationId": "PostController_publishPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostSchema"
                }
              }
            }
          }
        },
        "tags": ["Admin Posts"]
      }
    },
    "/api/admin/posts/{id}/unpublish": {
      "patch": {
        "operationId": "PostController_unpublishPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostSchema"
                }
              }
            }
          }
        },
        "tags": ["Admin Posts"]
      }
    },
...
    "/api/public/posts/{slug}": {
      "get": {
        "operationId": "PublicPostController_getPost",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostSchema"
                }
              }
            }
          }
        },
        "tags": ["Public Posts"]
      }
    },
    "/api/public/posts": {
      "get": {
        "operationId": "PublicPostController_getPosts",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Starting position of the query",
            "schema": {
              "default": 0,
              "description": "Starting position of the query",
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items to return",
            "schema": {
              "default": 20,
              "description": "Number of items to return",
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Schema for sorting items",
            "schema": {
              "$ref": "#/components/schemas/PublicPostController_getPosts_sortArray"
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filtering query string, in the format of \"property:rule[:value];property:rule[:value];...\"\nAvailable rules: eq, neq, gt, gte, lt, lte, like, nlike, in, nin, isnull, isnotnull \nAvailable properties: title, tag",
            "schema": {
              "$ref": "#/components/schemas/PublicPostController_getPosts_filterArray"
            },
            "style": "form",
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostsSchema"
                }
              }
            }
          }
        },
        "tags": ["Public Posts"]
      }
    },
    "/api/public/posts/{slug}/like": {
      "post": {
        "operationId": "PublicPostController_likePost",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostSchema"
                }
              }
            }
          }
        },
        "tags": ["Public Posts"]
      }
    },
    "/api/public/authors/{slug}/posts": {
      "get": {
        "operationId": "PublicAuthorController_getAuthorPosts",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Starting position of the query",
            "schema": {
              "default": 0,
              "description": "Starting position of the query",
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items to return",
            "schema": {
              "default": 20,
              "description": "Number of items to return",
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Schema for sorting items",
            "schema": {
              "$ref": "#/components/schemas/PublicAuthorController_getAuthorPosts_sortArray"
            },
            "style": "form",
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicAuthorPostsSchema"
                }
              }
            }
          }
        },
        "tags": ["Public Authors"]
      }
    },
...
  },
  "info": {
    "title": "Lonestone API",
    "description": "The Lonestone API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PostContentSchema": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": ["text"]
              },
              "data": {
                "type": "string"
              }
            },
            "required": ["type", "data"],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": ["image"]
              },
              "data": {
                "type": "string"
              }
            },
            "required": ["type", "data"],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": ["video"]
              },
              "data": {
                "type": "string"
              }
            },
            "required": ["type", "data"],
            "additionalProperties": false
          }
        ],
        "title": "PostContentSchema",
        "description": "Schema for content items (text, image, video)",
        "type": "object"
      },
...
      "PostVersionSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "title": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": ["id", "title", "createdAt"],
        "additionalProperties": false,
        "title": "PostVersionSchema",
        "description": "Schema for a post version"
      },
      "TagSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": ["id", "name", "slug"],
        "additionalProperties": false,
        "title": "TagSchema",
        "description": "A tag attached to a post"
      },
      "UserPostSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostContentSchema"
            }
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostVersionSchema"
            }
          },
          "publishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "type": "string",
            "enum": ["published", "draft"]
          },
          "commentCount": {
            "type": "number"
          },
          "coverImage": {
            "type": "string",
            "format": "uri"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagSchema"
            }
          }
        },
        "required": ["id", "title", "content", "versions", "type", "tags"],
        "additionalProperties": false,
        "title": "UserPostSchema",
        "description": "Schema for a user's post"
      },
...
      "UserPostsSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "slug": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "title": {
                  "type": "string"
                },
                "versions": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostVersionSchema"
                  }
                },
                "publishedAt": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": {
                  "type": "string",
                  "enum": ["published", "draft"]
                },
                "commentCount": {
                  "type": "number"
                },
                "coverImage": {
                  "type": "string",
                  "format": "uri"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagSchema"
                  }
                },
                "contentPreview": {
                  "$ref": "#/components/schemas/PostContentSchema"
                }
              },
              "required": ["id", "title", "versions", "type", "tags", "contentPreview"],
              "additionalProperties": false
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "offset": {
                "type": "number"
              },
              "pageSize": {
                "type": "number"
              },
              "itemCount": {
                "type": "number"
              },
              "hasMore": {
                "type": "boolean"
              }
            },
            "required": ["offset", "pageSize", "itemCount", "hasMore"],
            "additionalProperties": false
          }
        },
        "required": ["data", "meta"],
        "additionalProperties": false,
        "title": "UserPostsSchema",
        "description": "Schema for a list of user's posts"
      },
      "PublicPostSchema": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              }
            },
            "required": ["name"],
            "additionalProperties": false
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostContentSchema"
            }
          },
          "publishedAt": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "commentCount": {
            "type": "number"
          },
          "coverImage": {
            "type": "string",
            "format": "uri"
          },
          "likesCount": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagSchema"
            }
          }
        },
        "required": ["title", "author", "content", "publishedAt", "likesCount", "tags"],
        "additionalProperties": false,
        "title": "PublicPostSchema",
        "description": "A public post"
      },
      "PublicPostsSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "author": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": ["name"],
                  "additionalProperties": false
                },
                "publishedAt": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                },
                "commentCount": {
                  "type": "number"
                },
                "coverImage": {
                  "type": "string",
                  "format": "uri"
                },
                "likesCount": {
                  "type": "number"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagSchema"
                  }
                },
                "contentPreview": {
                  "$ref": "#/components/schemas/PostContentSchema"
                }
              },
              "required": [
                "title",
                "author",
                "publishedAt",
                "likesCount",
                "tags",
                "contentPreview"
              ],
              "additionalProperties": false
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "offset": {
                "type": "number"
              },
              "pageSize": {
                "type": "number"
              },
              "itemCount": {
                "type": "number"
              },
              "hasMore": {
                "type": "boolean"
              }
            },
            "required": ["offset", "pageSize", "itemCount", "hasMore"],
            "additionalProperties": false
          }
        },
        "required": ["data", "meta"],
        "additionalProperties": false,
        "title": "PublicPostsSchema",
        "description": "A list of public posts"
      },
      "PublicAuthorPostsSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "author": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": ["name"],
                  "additionalProperties": false
                },
                "publishedAt": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                },
                "commentCount": {
                  "type": "number"
                },
                "coverImage": {
                  "type": "string",
                  "format": "uri"
                },
                "likesCount": {
                  "type": "number"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagSchema"
                  }
                },
                "contentPreview": {
                  "$ref": "#/components/schemas/PostContentSchema"
                }
              },
              "required": [
                "title",
                "author",
                "publishedAt",
                "likesCount",
                "tags",
                "contentPreview"
              ],
              "additionalProperties": false
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "offset": {
                "type": "number"
              },
              "pageSize": {
                "type": "number"
              },
              "itemCount": {
                "type": "number"
              },
              "hasMore": {
                "type": "boolean"
              }
            },
            "required": ["offset", "pageSize", "itemCount", "hasMore"],
            "additionalProperties": false
          }
        },
        "required": ["data", "meta"],
        "additionalProperties": false,
        "title": "PublicAuthorPostsSchema",
        "description": "A list of posts from a specific author"
      },
...
      "PostController_getUserPosts_sortItem": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "enum": ["title", "createdAt"]
          },
          "direction": {
            "type": "string",
            "enum": ["asc", "desc"]
          }
        },
        "required": ["property", "direction"],
        "additionalProperties": false
      },
      "PostController_getUserPosts_sortArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PostController_getUserPosts_sortItem"
        }
      },
      "PostController_getUserPosts_filterItem": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "enum": ["title", "tag"]
          },
          "rule": {
            "type": "string",
            "enum": [
              "eq",
              "neq",
              "gt",
              "gte",
              "lt",
              "lte",
              "like",
              "nlike",
              "in",
              "nin",
              "isnull",
              "isnotnull"
            ]
          },
          "value": {
            "type": "string"
          }
        },
        "required": ["property", "rule"],
        "additionalProperties": false
      },
      "PostController_getUserPosts_filterArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PostController_getUserPosts_filterItem"
        }
      },
      "PublicPostController_getPosts_sortItem": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "enum": ["title", "createdAt"]
          },
          "direction": {
            "type": "string",
            "enum": ["asc", "desc"]
          }
        },
        "required": ["property", "direction"],
        "additionalProperties": false
      },
      "PublicPostController_getPosts_sortArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PublicPostController_getPosts_sortItem"
        }
      },
      "PublicPostController_getPosts_filterItem": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "enum": ["title", "tag"]
          },
          "rule": {
            "type": "string",
            "enum": [
              "eq",
              "neq",
              "gt",
              "gte",
              "lt",
              "lte",
              "like",
              "nlike",
              "in",
              "nin",
              "isnull",
              "isnotnull"
            ]
          },
          "value": {
            "type": "string"
          }
        },
        "required": ["property", "rule"],
        "additionalProperties": false
      },
      "PublicPostController_getPosts_filterArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PublicPostController_getPosts_filterItem"
        }
      },
...
      "PublicAuthorController_getAuthorPosts_sortArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PublicAuthorController_getAuthorPosts_sortItem"
        }
      }
    }
  }
}